requirements.txt
aiohttp==3.6.2
async-timeout==3.0.1
attrs==19.3.0
beautifulsoup4==4.8.2
certifi==2020.4.5
chardet==3.0.4
click==7.1.1
dataclasses==0.7
fastapi==0.54.0
future==0.18.2
h11==0.9.0
httptools==0.1.1
idna==2.9
idna-ssl==1.1.0
multidict==4.7.5
nltk==3.4.5
pkg-resources==0.0.0
pydantic==1.4
pyngrok==2.1.3
python-dotenv==0.12.0
PyYAML==5.3.1
requests==2.23.0
six==1.14.0
slackclient==2.5.0
soupsieve==2.0
starlette==0.13.2
typing-extensions==3.7.4.2
urllib3==1.25.8
uvicorn==0.11.3
uvloop==0.14.0
websockets==8.1
wikipedia==1.4.0
yarl==1.4.2
